# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowDesk API",
    "version": "1.0.0",
    "description": "FlowDesk secure multi-tenant customer engagement API."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development API Server"
    },
    {
      "url": "https://api.flowdesk.dev",
      "description": "Production API Server"
    }
  ],
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "summary": "List organizations for authenticated user",
        "operationId": "listUserOrganizations",
        "responses": {
          "200": {
            "description": "List of user organizations and active roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUserOrganizationsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Bootstrap a new organization",
        "operationId": "bootstrapOrganization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization successfully bootstrapped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapOrganizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid slug",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organizations/{orgId}/channels": {
      "get": {
        "summary": "List WhatsApp channels",
        "operationId": "listChannels",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant channels without stored credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Channel"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create a WhatsApp channel",
        "operationId": "createChannel",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Channel created; credentials are never returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid channel input"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/organizations/{orgId}/conversations": {
      "get": {
        "summary": "List conversations with filters and pagination",
        "operationId": "listConversations",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["new", "open", "pending", "resolved", "closed"]
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queueId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of conversations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListConversationsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": ["type", "title", "status", "code", "detail", "requestId"]
      },
...
      "Channel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["whatsapp"]
          },
          "name": {
            "type": "string"
          },
          "phoneNumberId": {
            "type": "string"
          },
          "wabaId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "connecting", "active", "degraded", "disconnected"]
          },
          "statusReason": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "type",
          "name",
          "phoneNumberId",
          "wabaId",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateChannelRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["whatsapp"],
            "default": "whatsapp"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "phoneNumberId": {
            "type": "string",
            "minLength": 1
          },
          "wabaId": {
            "type": "string",
            "minLength": 1
          },
          "accessToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8192,
            "writeOnly": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["name", "phoneNumberId", "wabaId", "accessToken"]
      },
...
      "BootstrapOrganizationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{1,62}$"
          }
        },
        "required": ["name", "slug"]
      },
      "UserOrganization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "membershipId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["id", "slug", "name", "role", "membershipId"]
      },
      "ListUserOrganizationsResponse": {
        "type": "object",
        "properties": {
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserOrganization"
            }
          }
        },
        "required": ["organizations"]
      },
      "BootstrapOrganizationResponse": {
        "type": "object",
        "properties": {
          "organization": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "slug": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              },
              "ownerRoleId": {
                "type": "string",
                "format": "uuid"
              },
              "membershipId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": ["id", "slug", "displayName", "ownerRoleId", "membershipId"]
          }
        },
        "required": ["organization"]
      },
...
      "Conversation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "customerPhone": {
            "type": "string"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["new", "open", "pending", "resolved", "closed"]
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "urgent"]
          },
          "assignedToUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "queueId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teamId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "waitingReason": {
            "type": "string",
            "nullable": true
          },
          "botPaused": {
            "type": "boolean"
          },
          "firstResponseDueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolutionDueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "firstRespondedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "slaPausedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "firstResponseRemainingSeconds": {
            "type": "integer",
            "minimum": 0,
            "nullable": true
          },
          "resolutionRemainingSeconds": {
            "type": "integer",
            "minimum": 0,
            "nullable": true
          },
          "version": {
            "type": "integer"
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "channelId",
          "customerPhone",
          "customerName",
          "status",
          "priority",
          "assignedToUserId",
          "queueId",
          "teamId",
          "waitingReason",
          "botPaused",
          "firstResponseDueAt",
          "resolutionDueAt",
          "resolvedAt",
          "firstRespondedAt",
          "slaPausedAt",
          "firstResponseRemainingSeconds",
          "resolutionRemainingSeconds",
          "version",
          "lastMessageAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ListConversationsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Conversation"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["items", "nextCursor"]
      },
...
    }
  }
}